requirements.txt
discord.py
httpx
openai
pyyaml
sentence-transformers
scikit-learn
numpy
neo4j